+++ __init__.py
@@ -26,43 +26,126 @@ bl_info = {
 }
 
 import bpy
+import os
 import addon_utils
+
+if 'bpy' in locals():
+    if 'PreflightPanel' in locals():
+        print("Reload Event Detected...")
+        import imp
+        imp.reload(properties)
+        imp.reload(operators)
+        imp.reload(ui)
 
 from .properties import *
 from .operators import *
-from .preflight import PreflightPanel
+from .ui import *
 
-# Reload all modules because blender doesn't do this.
-# Sourced from: https://github.com/zeffii/mesh_tiny_cad
-if "bpy" in locals():
-    if 'Preflight' in locals():
+LARGE_BUTTON_SCALE_Y = 1.5
 
-        print('Preflight: detected reload event.')
-        import importlib
 
-        try:
-            modules = (properties, operators, preflight)
-            for m in modules:
-                importlib.reload(m)
-            print("Preflight: reloaded modules, all systems operational")
+class PreflightPanel(bpy.types.Panel):
+    bl_space_type = "VIEW_3D"
+    bl_region_type = "TOOLS"
+    bl_label = "Pre-Flight FBX"
+    bl_context = "objectmode"
+    bl_category = "Pre-Flight FBX"
 
-        except Exception as E:
-            print('reload failed with error:')
-            print(E)
+    def draw(self, context):
+        layout = self.layout
+        groups = context.scene.preflight_props.fbx_export_groups
+
+        # Export Groups
+        layout.operator("preflight.add_export_group",
+                        text="Add Export Group", icon="ZOOMIN")
+        for group_idx, group in enumerate(groups):
+            self.layout_export_group(group_idx, group, layout, context)
+
+        layout.separator()
+
+        layout.prop(context.scene.preflight_props, "export_location",
+                    icon="LIBRARY_DATA_DIRECT", text="")
+        layout.prop(context.scene.preflight_props, "export_animations")
+
+        layout.separator()
+
+        # Export Button
+        export_row = layout.row()
+        export_row.scale_y = LARGE_BUTTON_SCALE_Y
+        exportButton = export_row.operator(
+            "preflight.export_groups",
+            icon="EXPORT")
+
+    def layout_export_group(self, group_idx, group, layout, context):
+        group_box = layout.box()
+
+        # Header Row
+        row = group_box.row()
+
+        row.prop(
+            group,
+            "is_collapsed",
+            icon="TRIA_RIGHT" if group.is_collapsed else "TRIA_DOWN",
+            icon_only=True,
+            emboss=False
+        )
+
+        row.prop(group, "name", text="")
+
+        remove_group_button = row.operator(
+            "preflight.remove_export_group",
+            text="",
+            icon="X"
+        )
+
+        remove_group_button.group_idx = group_idx
+
+        if group.is_collapsed is False:
+            # Mesh Collection
+            group_box.label("Objects to Export:",
+                            icon="OUTLINER_OB_GROUP_INSTANCE")
+            mesh_column = group_box.column(align=True)
+            for mesh_idx, mesh in enumerate(group.obj_names):
+                self.layout_mesh_row(mesh_idx, group_idx,
+                                     mesh, mesh_column, context)
+
+            # Add Mesh Button
+            add_mesh_button = mesh_column.operator(
+                "preflight.add_object_to_group", text="Add Object", icon="ZOOMIN")
+            add_mesh_button.group_idx = group_idx
+
+            group_box.template_list(
+                "ExportObjectUIList", "obj_list", group, "obj_names", group, "obj_idx")
+
+            # Export Options
+            group_box.separator()
+            options_column = group_box.column(align=True)
+            options_column.prop(group, "include_armatures")
+            options_column.prop(group, "include_animations")
+            options_column.prop(group, "apply_modifiers")
+
+    def layout_mesh_row(self, mesh_idx, group_idx, mesh, layout, context):
+        mesh_row = layout.row(align=True)
+        mesh_row.prop_search(mesh, "obj_name", context.scene,
+                             "objects", text="", icon="OBJECT_DATA")
+
+        # Remove Mesh Button
+        removeMeshButton = mesh_row.operator(
+            "preflight.remove_object_from_group", text="", icon="ZOOMOUT")
+        removeMeshButton.group_idx = group_idx
+        removeMeshButton.object_idx = mesh_idx
 
 
 def register():
     bpy.utils.register_module(__name__)
     addon_utils.enable("io_scene_fbx", default_set=True, persistent=True)
     bpy.types.Scene.preflight_props = bpy.props.PointerProperty(
-        name="Preflight Properties",
-        type=PreflightOptionsGroup
-    )
+        type=PreflightOptionsGroup)
 
 
 def unregister():
-    bpy.utils.unregister_module(__name__)
     del bpy.types.Scene.preflight_props
+    bpy.utils.unregister_module(__name__)
 
 
 if __name__ == "__main__":
